feat: separate test for modules
