+++ .env.example
@@ -8,4 +8,4 @@ DB_HOST=localhost
 DB_PORT=3333
 DB_USER=test
 DB_PASSWORD=test
-DB_DATABASE=db
+DB_DATABASE=db.sqlite
