+++ .eslintrc.json
@@ -14,6 +14,16 @@
     "plugins": [
         "@typescript-eslint"
     ],
-    "rules": {
-    }
+    "rules": {},
+    "overrides": [
+        {
+            "files": [
+                "**/*.spec.ts",
+                "jest/*.ts"
+            ],
+            "env": {
+                "jest": true
+            }
+        }
+    ]
 }
