+++ ormconfig.js
@@ -15,9 +15,6 @@ module.exports = {
   entities: [
     process.env.TYPEORM_ENTITIES
   ],
-  migrations: [
-    'src/database/migrations/*.ts'
-  ],
   keepConnectionAlive: true,
   synchronize: true
 }
