+++ package.json
@@ -12,7 +12,7 @@
     "start": "node dist/index.js",
     "build": "babel src --extensions \".js,.ts\" --out-dir dist --copy-files --no-copy-ignored",
     "dev": "ts-node-dev --respawn --transpile-only --ignore-watch node_modules src/index.ts",
-    "test": "jest --coverage"
+    "test": "jest --coverage --runInBand"
   },
   "devDependencies": {
     "@babel/cli": "^7.13.16",
@@ -23,6 +23,7 @@
     "@babel/preset-typescript": "^7.13.0",
     "@types/express": "^4.17.11",
     "@types/node": "^15.0.2",
+    "@types/supertest": "^2.0.11",
     "@typescript-eslint/eslint-plugin": "^4.22.1",
     "@typescript-eslint/parser": "^4.22.1",
     "babel-plugin-module-resolver": "^4.1.0",
